backend/requirements.txt
# Core Framework
fastapi
uvicorn
python-dotenv
pydantic
python-multipart

# LangChain Ecosystem (Pinned for stability)
langchain~=0.3.0
langchain-ollama
langchain-community
langchain-chroma

# Data & Extraction
chromadb
sentence-transformers
pypdf
requests
beautifulsoup4
lxml
